move utils where they belong
